AlphaFold_demo/scripts/model.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-

import os
import numpy as np
import tensorflow as tf
from pathlib import Path
import yaml
import time
from Bio import SeqIO
from Bio.PDB import PDBParser, PDBIO
from Bio.PDB.Structure import Structure
from Bio.PDB.Model import Model
from Bio.PDB.Chain import Chain
from Bio.PDB.Residue import Residue
from Bio.PDB.Atom import Atom

# 标准氨基酸三字母缩写映射
amino_acids = {
    'A': 'ALA', 'R': 'ARG', 'N': 'ASN', 'D': 'ASP', 'C': 'CYS', 
    'Q': 'GLN', 'E': 'GLU', 'G': 'GLY', 'H': 'HIS', 'I': 'ILE', 
    'L': 'LEU', 'K': 'LYS', 'M': 'MET', 'F': 'PHE', 'P': 'PRO', 
    'S': 'SER', 'T': 'THR', 'W': 'TRP', 'Y': 'TYR', 'V': 'VAL'
}

class SimplifiedAlphaFold:
    """简化的AlphaFold模型，用于演示"""
    
    def __init__(self, config_path=None):
        """初始化模型"""
        self.config = {}
        if config_path:
            self.load_config(config_path)
        
        # 氨基酸及其对应编号
        self.aa_types = {
            'A': 0, 'R': 1, 'N': 2, 'D': 3, 'C': 4, 'Q': 5, 'E': 6, 
            'G': 7, 'H': 8, 'I': 9, 'L': 10, 'K': 11, 'M': 12, 'F': 13, 
            'P': 14, 'S': 15, 'T': 16, 'W': 17, 'Y': 18, 'V': 19
        }
        
        # 简单的原子坐标映射（简化版，实际上需要更精确的模型）
        self.atom_positions = {aa: np.random.rand(4, 3) * 5 for aa in self.aa_types}
        
        # 初始化模型状态
        self.is_ready = False
        try:
            print("加载模型...")
            time.sleep(1)  # 模拟加载时间
            self.is_ready = True
            print("模型加载完成")
        except Exception as e:
            print(f"模型加载失败: {e}")
            raise
    
    def load_config(self, config_path):
        """加载模型配置"""
        try:
            with open(config_path, 'r') as f:
                self.config = yaml.safe_load(f)
            
            # 验证配置文件
            required_sections = ['model', 'inference', 'data']
            for section in required_sections:
                if section not in self.config:
                    raise ValueError(f"配置文件缺少必要的部分: {section}")
            
            # 设置模型参数
            self.max_seq_length = self.config['inference']['max_seq_length']
            self.num_ensemble = self.config['inference']['num_ensemble']
            self.num_recycle = self.config['inference']['num_recycle']
            self.model_preset = self.config['inference']['model_preset']
            
            print(f"已加载配置:")
            print(f"- 模型名称: {self.config['model']['name']}")
            print(f"- 最大序列长度: {self.max_seq_length}")
            print(f"- 集成数量: {self.num_ensemble}")
            print(f"- 循环次数: {self.num_recycle}")
            print(f"- 模型预设: {self.model_preset}")
            
        except Exception as e:
            print(f"加载配置文件失败: {e}")
            raise RuntimeError(f"配置文件加载失败: {str(e)}")
    
    def preprocess(self, sequence):
        """预处理蛋白质序列"""
        print(f"预处理序列: {sequence[:10]}...")
        
        # 将氨基酸序列转换为数字编码
        seq_encoded = np.array([self.aa_types.get(aa, -1) for aa in sequence])
        
        # 检查无效字符
        if -1 in seq_encoded:
            invalid_chars = [aa for aa in sequence if aa not in self.aa_types]
            print(f"警告: 序列中包含无效字符: {invalid_chars}")
            seq_encoded = np.array([self.aa_types.get(aa, 0) for aa in sequence])
        
        # 创建一个简单的特征输入（实际上AlphaFold会有更复杂的特征提取）
        seq_length = len(sequence)
        features = {
            'sequence': sequence,
            'seq_length': seq_length,
            'seq_encoded': seq_encoded,
            'msa_features': np.random.rand(1, seq_length, 20)  # 假装有MSA特征
        }
        
        return features
    
    def predict(self, sequence, output_dir):
        """预测蛋白质结构"""
        if not self.is_ready:
            raise RuntimeError("模型未正确加载")
        
        if not sequence:
            raise ValueError("序列不能为空")
        
        try:
            # 预处理
            features = self.preprocess(sequence)
            
            print("开始预测蛋白质结构...")
            time.sleep(2)  # 模拟预测时间
            
            # 构建一个简单的结构输出
            seq_length = features['seq_length']
            
            # 创建简单的坐标（真实的AlphaFold会生成精确的3D坐标）
            # 这里我们生成一个简单的螺旋结构
            t = np.linspace(0, 10 * np.pi, seq_length)
            x = np.cos(t) * 5
            y = np.sin(t) * 5
            z = t
            
            # 组合成原子坐标
            atom_positions = []
            for i, aa in enumerate(sequence):
                # 每个氨基酸4个主要原子: N, CA, C, O
                base_pos = np.array([x[i], y[i], z[i]])
                
                # 使用预定义的原子相对位置
                aa_idx = self.aa_types.get(aa, 0)
                for atom_offset in self.atom_positions[aa]:
                    atom_positions.append(base_pos + atom_offset)
            
            atom_positions = np.array(atom_positions).reshape(-1, 3)
            
            # 计算置信度分数（模拟的）
            confidence = np.linspace(0.5, 0.9, seq_length)
            
            # 创建结果字典
            result = {
                'atom_positions': atom_positions,
                'sequence': sequence,
                'confidence': confidence,
                'plddt': np.random.rand(seq_length) * 70 + 30,  # 模拟pLDDT得分
            }
            
            # 确保输出目录存在
            os.makedirs(output_dir, exist_ok=True)
            
            # 将结果保存为PDB文件
            pdb_path = self.save_pdb(result, output_dir)
            
            if not os.path.exists(pdb_path):
                raise RuntimeError("PDB文件生成失败")
            
            print(f"预测完成，结果保存到: {pdb_path}")
            return result, pdb_path
            
        except Exception as e:
            print(f"预测过程出错: {e}")
            raise RuntimeError(f"预测失败: {str(e)}")
    
    def save_pdb(self, result, output_dir):
        """将预测结果保存为PDB格式文件"""
        try:
            os.makedirs(output_dir, exist_ok=True)
            pdb_path = os.path.join(output_dir, "predicted_structure.pdb")
            
            # 创建PDB结构
            structure = Structure("predicted")
            model = Model(0)
            structure.add(model)
            chain = Chain("A")
            model.add(chain)
            
            # 添加残基和原子
            atom_index = 0
            atom_positions = result['atom_positions'].reshape(-1, 3)
            
            for i, aa in enumerate(result['sequence']):
                # 创建残基
                res_name = amino_acids.get(aa, "UNK")
                res = Residue((" ", i+1, " "), res_name, " ")
                chain.add(res)
                
                # 添加4个主要原子: N, CA, C, O
                atom_names = ["N", "CA", "C", "O"]
                for j, atom_name in enumerate(atom_names):
                    if atom_index < len(atom_positions):
                        coord = atom_positions[atom_index]
                        atom = Atom(
                            name=atom_name,
                            coord=coord,
                            bfactor=float(result['plddt'][i]),  # 使用pLDDT作为B-factor
                            occupancy=1.0,
                            altloc=" ",
                            fullname=atom_name,
                            serial_number=atom_index+1,
                            element=atom_name[0]
                        )
                        res.add(atom)
                        atom_index += 1
            
            # 保存PDB文件
            io = PDBIO()
            io.set_structure(structure)
            io.save(pdb_path)
            
            return pdb_path
            
        except Exception as e:
            print(f"保存PDB文件时出错: {e}")
            raise RuntimeError(f"保存PDB文件失败: {str(e)}") 