AlphaFold_demo/requirements.txt
numpy>=1.19.5
tensorflow>=2.5.0
biopython>=1.79
matplotlib>=3.4.3
py3Dmol>=1.8.0
requests>=2.25.1
tqdm>=4.60.0
pyyaml>=5.4.1
pandas>=1.3.0
scikit-learn>=0.24.2
seaborn>=0.11.1
jupyter>=1.0.0
ipywidgets>=7.6.3 